Moved tests to test_typeconversions.
