+++ src/tests/test_typeconversions.py
@@ -271,3 +271,38 @@ def test_jsproxy_attribute_error(selenium):
     with pytest.raises(WebDriverException, match=msg):
         selenium.run("point.y")
     assert selenium.run_js("return point.y;") is None
+
+def test_javascript_error(selenium):
+    msg = "JsException: Error: This is a js error"
+    with pytest.raises(WebDriverException, match=msg):
+        selenium.run(
+            """
+            from js import Error
+            err = Error.new("This is a js error")
+            err2 = Error.new("This is another js error")
+            raise err
+            """
+        )
+
+
+def test_javascript_error_back_to_js(selenium):
+    selenium.run_js(
+        """
+        window.err = new Error("This is a js error")
+        """
+    )
+    assert (
+        selenium.run(
+            """
+        from js import err
+        py_err = err
+        type(py_err).__name__
+        """
+        )
+        == "JsException"
+    )
+    assert selenium.run_js(
+        """
+        return pyodide.globals["py_err"] === err
+        """
+    )
